convert `blitz seed` to `blitz db seed`
