+++ packages/cli/src/commands/db.ts
@@ -1,5 +1,9 @@
 import {Command, flags} from "@oclif/command"
 import {log} from "@blitzjs/display"
+import {join} from "path";
+import pkgDir from "pkg-dir";
+
+const projectRoot = pkgDir.sync() || process.cwd()
 
 const getPrismaBin = () => require("@blitzjs/server").resolveBinAsync("@prisma/cli", "prisma")
 let prismaBin: string
@@ -148,6 +152,41 @@ export function getDbName(connectionString: string): string {
   return dbName
 }
 
+async function runSeed() {
+  log.branded("Seeding database")
+  let spinner = log.spinner("Loading seeds").start()
+
+  let seeds: Function
+
+  try {
+    seeds = require(join(projectRoot, "db/seeds")).default
+
+    if (seeds === undefined) {
+      throw new Error(`Cant find default export from db/seeds`)
+    }
+  } catch (err) {
+    log.error(err)
+    log.error(`Couldn't import default from db/seeds.ts or db/seeds/index.ts file`)
+  }
+  spinner.succeed()
+
+  spinner = log.spinner("Checking for database migrations\n").start()
+  await runMigrate({}, `--schema=${join(process.cwd(), "db", "schema.prisma")}`)
+  spinner.succeed()
+
+  try {
+    console.log(log.withCaret("Seeding..."))
+    await seeds()
+  } catch (err) {
+    log.error(err)
+    log.error(`Couldn't run imported function, are you sure it's a function?`)
+  }
+
+  const db = require(join(projectRoot, "db/index")).default
+  await db.disconnect()
+  log.success("Done seeding")
+}
+
 export class Db extends Command {
   static description = `Run database commands
 
@@ -166,6 +205,10 @@ ${require("chalk").bold(
 ${require("chalk").bold(
   "reset",
 )}   Reset the database and run a fresh migration via Prisma 2. You can also pass --force to skip all the user prompts.
+
+${require("chalk").bold(
+  "seed",
+)}   Generates seeded data in database via Prisma 2. You need db/seeds.ts or db/seeds/index.ts.
 `
 
   static args = [
@@ -260,6 +303,10 @@ ${require("chalk").bold(
 
     if (command === "help") {
       return Db.run(["--help"])
+    }
+
+    if (command === "seed") {
+      return await runSeed()
     }
 
     this.log("\nUh oh, Blitz does not support that command.")
